Added club and team classes
Added club and team classes. There are 4 types of teams, Men, Men U19, Female and Female U19. When choosing a country, it now presents Men teams from that country. The goal is to present the leagues for each country and then choose from that league and then be able to choose both womens teams and mens teams.